requirements/dev.txt
-r common.txt
coverage==7.6.10
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0